Abstract store
